chore(cli): refactor CLI authentication flow and tests
- move handleLogin, handleLogout, and handleStatus functions to cli-functions.ts for better separation of concerns.
- remove deprecated login method from AuthService and streamline the login process using loginWithCallback.
- remove redundant tests.
