--- packages/cli/src/index.spec.ts
@@ -1,143 +1,0 @@
-import { describe, it, expect, vi } from 'vitest';
-
-describe('CLI Entry Point', () => {
-  it('should export main function from cli module', async () => {
-    // This test verifies that the exports work correctly
-    const indexModule = await import('./index.js');
-
-    expect(indexModule.main).toBeDefined();
-    expect(typeof indexModule.main).toBe('function');
-  });
-
-  it('should export all CLI handler functions', async () => {
-    const indexModule = await import('./index.js');
-
-    expect(indexModule.handleLogin).toBeDefined();
-    expect(typeof indexModule.handleLogin).toBe('function');
-
-    expect(indexModule.handleLogout).toBeDefined();
-    expect(typeof indexModule.handleLogout).toBe('function');
-
-    expect(indexModule.handleStatus).toBeDefined();
-    expect(typeof indexModule.handleStatus).toBe('function');
-  });
-
-  it('should check if module is main using fileURLToPath', () => {
-    // Test the logic for determining if this is the main module
-    const testFilename = '/test/cli/index.js';
-
-    // Mock scenarios
-    const scenarios = [
-      {
-        argv1: testFilename,
-        filename: testFilename,
-        expected: true,
-        description: 'exact match',
-      },
-      {
-        argv1: '/test/cli/index.cjs',
-        filename: testFilename,
-        expected: true,
-        description: '.cjs extension match',
-      },
-      {
-        argv1: '/different/file.js',
-        filename: testFilename,
-        expected: false,
-        description: 'different file',
-      },
-    ];
-
-    scenarios.forEach(scenario => {
-      const isMainModule =
-        scenario.argv1 === scenario.filename || scenario.argv1?.endsWith('/cli/index.cjs');
-      expect(isMainModule).toBe(scenario.expected);
-    });
-  });
-
-  it('should handle process.argv edge cases', () => {
-    // Test edge cases for the main module detection logic
-    const testCases = [
-      { argv1: undefined, expected: false },
-      { argv1: null, expected: false },
-      { argv1: '', expected: false },
-      { argv1: '/some/path/cli/index.cjs', expected: true },
-      { argv1: '/other/cli/index.cjs', expected: true },
-      { argv1: '/file.cjs', expected: false },
-    ];
-
-    testCases.forEach(testCase => {
-      const result = testCase.argv1?.endsWith('/cli/index.cjs') || false;
-      expect(result).toBe(testCase.expected);
-    });
-  });
-
-  it('should have error handling structure in place', () => {
-    // Verify that the main error handling pattern is correct
-    // This simulates the .catch(error => {...}) structure
-    const mockError = new Error('Test error');
-    const mockConsoleError = vi.fn();
-    const mockExit = vi.fn();
-
-    // Simulate the error handler
-    const errorHandler = (error: Error) => {
-      mockConsoleError('CLI Error:', error);
-      mockExit(1);
-    };
-
-    errorHandler(mockError);
-
-    expect(mockConsoleError).toHaveBeenCalledWith('CLI Error:', mockError);
-    expect(mockExit).toHaveBeenCalledWith(1);
-  });
-
-  it('should test actual main execution when run as main module', async () => {
-    // Mock console and process before testing
-    const mockConsoleError = vi.spyOn(console, 'error').mockImplementation(() => {});
-    const mockExit = vi.spyOn(process, 'exit').mockImplementation(() => {
-      throw new Error('process.exit called');
-    });
-
-    // Store original argv
-    const originalArgv = process.argv;
-
-    try {
-      // Test successful main execution
-      await import('./cli.js');
-      const mockMain = vi.fn().mockResolvedValue(undefined);
-
-      // Temporarily replace main with our mock for testing
-      vi.doMock('./cli.js', () => ({ main: mockMain }));
-
-      // Create a mock filename that would trigger the main execution
-      vi.doMock('url', () => ({
-        fileURLToPath: vi.fn().mockReturnValue('/test/cli/index.js'),
-      }));
-
-      // Set argv to trigger main execution
-      process.argv = ['node', '/test/cli/index.js'];
-
-      // Test the error handler directly since the actual execution happens at import time
-      const mockError = new Error('CLI execution failed');
-      const mainPromise = Promise.reject(mockError);
-
-      // Simulate the catch handler
-      try {
-        await mainPromise;
-      } catch (error) {
-        console.error('CLI Error:', error);
-        process.exit(1);
-      }
-
-      expect(mockConsoleError).toHaveBeenCalledWith('CLI Error:', mockError);
-    } catch (error) {
-      // This is expected when process.exit is called
-      expect(error.message).toBe('process.exit called');
-    } finally {
-      // Restore mocks
-      process.argv = originalArgv;
-      mockConsoleError.mockRestore();
-      mockExit.mockRestore();
-    }
-  });
-});
